+++ wazo_agentd/service/handler/agent_queues.py
@@ -22,6 +22,13 @@ class AgentQueuesHandler:
             )
         return self._handle_list_user_queues(agent)
 
+    @debug.trace_duration
+    def handle_list_queues_by_id(self, agent_id, tenant_uuids=None):
+        logger.info('Executing list_queues_by_id command (agent %s)', agent_id)
+        with db_utils.session_scope():
+            agent = self._agent_dao.get_agent(agent_id, tenant_uuids=tenant_uuids)
+        return self._handle_list_user_queues(agent)
+
     def _handle_list_user_queues(self, agent):
         return [
             {